+++ include/vinecopulib/vinecop/fit_controls.hpp
@@ -56,8 +56,9 @@ public:
     bool select_threshold = false,
     bool select_families = true,
     bool show_trace = false,
-    size_t num_threads = 1,
     std::string mst_algorithm = "prim");
+
+  explicit FitControlsVinecop(const FitControlsConfig& config);
 
   // Getters
   DEPRECATED size_t get_truncation_level() const;