+++ test/src_test/include/test_vinecop_class.hpp
@@ -565,6 +565,7 @@ TEST_F(VinecopTest, select_finds_right_structure_kruskal)
   // check whether the same structure appears if we only allow for
   // independence (pair-copula estimates differ otherwise)
   FitControlsVinecop controls({ BicopFamily::indep });
+  EXPECT_EQ(controls.get_mst_algorithm(), "prim");
   EXPECT_ANY_THROW(controls.set_mst_algorithm("foobar"));
   controls.set_mst_algorithm("kruskal");